requirements.txt
discord
mwclient
html2markdown
python-dotenv